메인 구현중

+++ jblog03/src/main/java/com/douzone/jblog/repository/CategoryRepository.java
@@ -39,7 +39,7 @@ public class CategoryRepository {
 		return sqlSession.delete("category.delete", no);
 	}
 
-	public int getFirstCategoryNo(String id) {
+	public Long getFirstCategoryNo(String id) {
 		return sqlSession.selectOne("category.firstCN", id);
 	}
